Move prepare uniform superposition to state_preparation subpackage
