+++ dev_tools/autogenerate-bloqs-notebooks-v2.py
@@ -66,12 +66,11 @@ import qualtran.bloqs.chemistry.trotter.grid_ham.inverse_sqrt
 import qualtran.bloqs.chemistry.trotter.grid_ham.qvr
 import qualtran.bloqs.factoring.mod_exp
 import qualtran.bloqs.mcmt.and_bloq
-import qualtran.bloqs.mcmt.multi_control_multi_target_pauli
-import qualtran.bloqs.prepare_uniform_superposition
 import qualtran.bloqs.qrom
 import qualtran.bloqs.reflection
 import qualtran.bloqs.rotations.phasing_via_cost_function
 import qualtran.bloqs.rotations.quantum_variable_rotation
+import qualtran.bloqs.state_preparation.prepare_uniform_superposition
 import qualtran.bloqs.state_preparation.state_preparation_via_rotation
 import qualtran.bloqs.swap_network
 
@@ -112,8 +111,10 @@ NOTEBOOK_SPECS: List[NotebookSpecV2] = [
     ),
     NotebookSpecV2(
         title='Prepare Uniform Superposition',
-        module=qualtran.bloqs.prepare_uniform_superposition,
-        bloq_specs=[qualtran.bloqs.prepare_uniform_superposition._PREP_UNIFORM_DOC],
+        module=qualtran.bloqs.state_preparation.prepare_uniform_superposition,
+        bloq_specs=[
+            qualtran.bloqs.state_preparation.prepare_uniform_superposition._PREP_UNIFORM_DOC
+        ],
         directory=f'{SOURCE_DIR}/bloqs/',
     ),
     NotebookSpecV2(
