+++ qualtran/bloqs/chemistry/sparse/sparse_test.py
@@ -19,7 +19,9 @@ from openfermion.resource_estimates.utils import power_two, QI
 
 from qualtran.bloqs.basic_gates import TGate
 from qualtran.bloqs.chemistry.sparse import PrepareSparse, SelectSparse
-from qualtran.bloqs.prepare_uniform_superposition import PrepareUniformSuperposition
+from qualtran.bloqs.state_preparation.prepare_uniform_superposition import (
+    PrepareUniformSuperposition,
+)
 from qualtran.testing import execute_notebook
