+++ qualtran/bloqs/state_preparation/state_preparation_alias_sampling.py
@@ -32,9 +32,11 @@ from qualtran import BoundedQUInt, Register, Signature
 from qualtran._infra.gate_with_registers import total_bits
 from qualtran.bloqs.arithmetic import LessThanEqual
 from qualtran.bloqs.basic_gates.swap import CSwap
-from qualtran.bloqs.prepare_uniform_superposition import PrepareUniformSuperposition
 from qualtran.bloqs.qrom import QROM
 from qualtran.bloqs.select_and_prepare import PrepareOracle
+from qualtran.bloqs.state_preparation.prepare_uniform_superposition import (
+    PrepareUniformSuperposition,
+)
 from qualtran.linalg.lcu_util import preprocess_lcu_coefficients_for_reversible_sampling
